+++ addons/html_builder/models/__init__.py
@@ -1,0 +1,1 @@
+from . import ir_qweb
